apps/api-gateway/tests/integration/test_api_gateway_integration.py
"""
Integration tests for Api-Gateway service.
"""

import pytest
import asyncio


class TestApigatewayIntegration:
    """Integration test suite for api-gateway service."""
    
    @pytest.mark.asyncio
    async def test_database_connection(self):
        """Test database connectivity."""
        # TODO: Implement database connection test
        pass
    
    @pytest.mark.asyncio
    async def test_redis_connection(self):
        """Test Redis connectivity."""
        # TODO: Implement Redis connection test
        pass
