apps/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
redis[hiredis]==5.0.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-auto-instrumentation==0.42b0
prometheus-client==0.19.0
